Do not pin the ”major” version of pycountry (#132)
Since pycountry uses a calendar-based versioning scheme, this doesn’t
have the effect it would for a SemVer dependency. Breaking changes are
not more likely to happen at the beginning of a calendar year than in
the middle of one, and pinning the version is likely to cause more
headaches than it prevents.